load models before collections
since an object literal form is used to store these models (M.xxx),
all these are loaded upfront. if we were to use `Bill` over `M.Bill`,
this would be loaded when we call it (lazy loaded) and by that time we
would have the models loaded completely on the page. since we are using
a single object `M` to hold all the models and collections, the order of
loading varies and when we define a collection, the model attribute would
be undefined.

M.Bills = Backbone.Collection.extend({
	model: M.Bill
});

If M.Bill wasn't loaded, this would be

M.Bills = Backbone.Collection.extend({
	model: undefined
});

and we get errors fetching the records from the server